+++ packages/types/CHANGELOG.md
@@ -2,7 +2,13 @@
 
 ## Unreleased
 
+<<<<<<< Updated upstream
 - Add support to page coverage by adding `pageId` to `Eyes.check`.
+=======
+>>>>>>> Stashed changes
+- add `waitForSelector` to `SpecDriver` type
+- add `WaitOptions` type
+
 ## 1.0.25 - 2022/2/15
 
 - add `viewportScale` property to `DriverInfo` type